ref: add type WalletAnalyzeResult
